requirements.txt
Flask==0.10.1
Jinja2==2.7.3
MarkupSafe==0.23
Werkzeug==0.10.1
argparse==1.2.1
httplib2==0.9
itsdangerous==0.24
oauth2==1.5.211
oauthlib==0.7.2
python-linkedin==4.1
requests==2.5.1
requests-oauthlib==0.4.2
simplejson==3.6.5
wsgiref==0.1.2
